requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.0.32
colorama==0.4.6